change length id ufs

+++ database/migrations/2020_11_09_005202_tbl_enderecos.php
@@ -22,7 +22,7 @@ class TblEnderecos extends Migration
             $table->string('complemento', 25)->nullable();
             $table->string('cidade', 40);
             $table->string('bairro', 40);
-            $table->unsignedTinyInteger('id_uf');
+            $table->unsignedSmallInteger('id_uf');
             $table->timestamps();
         });
     }
